fixed reset upon refresh, fixed vendor post requests upon signup, updated styling
