+++ ui/src/components/LandingPage/LandingPage.css
@@ -1,9 +1,12 @@
-#root {
-  /* max-width: 1280px; */
+/* #root {
+  max-width: 1280px;
   margin: 0 auto;
-  padding: 1rem;
   height: 100%;
-}
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+} */
+
 .landing h1 {
   color: black;
   font-family: "DM Serif Display", serif;
