+++ ui/src/components/Navbar/ShopperNavbar/ShopperNavbar.css
@@ -15,7 +15,7 @@
   padding: 0 80px;
 }
 /*container inside the navbar for the links*/
-.container {
+/* .container {
   max-width: 100%;
   width: 100%;
   margin: 0 auto;
@@ -23,7 +23,7 @@
   display: flex;
   flex-direction: row;
   align-items: center;
-}
+} */
 
 .links {
   display: flex;
@@ -36,8 +36,7 @@
   max-width: 650px;
   gap: 30px;
   transform: translateX(-130px);
-  color:black;
-  
+  color: black;
 }
 
 .account:hover {
