+++ ui/src/components/Product/ProductCard.css
@@ -32,7 +32,8 @@ a {
 
 .product-tumb img {
   max-width: 100%;
-  max-height: 100%;
+  max-height: 300px;
+  height: auto;
 }
 
 .product-details {
