+++ ui/src/components/Shoppingcart/Shoppingcart.css
@@ -7,6 +7,10 @@
   transform: translateY(30px);
 }
 
+.cart-banner h1 {
+  font-family: "DM Serif Display", serif;
+}
+
 .cart-page {
   display: flex;
   flex-direction: row;
@@ -21,7 +25,7 @@
 .shopping-cart {
   width: 800px;
   height: auto;
-  margin: 100px 20px 0 auto;
+  margin: 80px 20px 0 auto;
   background: #ffffff;
   box-shadow: 2px 2px 2px 2px rgba(0, 0, 0, 0.1);
   border-radius: 6px;
@@ -33,7 +37,7 @@
 .shopping-cart-totals {
   width: 600px;
   float: right;
-  margin: 100px auto 0 0;
+  margin: 80px auto 0 0;
   background: #ffffff;
   box-shadow: 2px 2px 2px 2px rgba(0.1, 0.1, 0.1, 0.1);
   border-radius: 6px;
