+++ ui/src/components/Store/StoreCard.css
@@ -4,6 +4,7 @@
   margin-left: auto;
   margin-right: auto;
   transform: translateY(120px);
+  margin-bottom: 30px;
 }
 
 .store-card-spacing {
