+++ ui/src/index.css
@@ -3,6 +3,16 @@
   font-family: "Roboto", sans-serif;
 }
 
+#root {
+  /* max-width: 1280px; */
+  /* margin: 0 auto; */
+  /* padding: 1rem; */
+  /* height: 100%; */
+  display: flex;
+  flex-direction: column;
+  flex-wrap: nowrap;
+}
+
 :root {
   font-size: 16px;
   line-height: 24px;
